HillClimberV2
The hill climber now clones the cycle plans to ensure that the mutations
are accurate based on the current best seed.
Also there is now a function to ensure that all stages are called in the
final plan.
There is a single stage generator which is based on the the delay per
second (which is the total delay divided by the total time needed for
that stage and inter-green)